[FEATURE/REFACTORING][Option(Name/PhoneNum/Birthday) 구현 및refactoring] - Option(Name/PhoneNum/Birthday) 구현 - 위임받은 객체에서 최소한의 동작으로 변경 - 상위 class에서 dataManager 참조하여 동작

+++ ClarifyPJT/ClarifyPJT/DataManager.h
@@ -18,7 +18,7 @@ public:
 
 	int getSize(void)
 	{
-		return _employees.size();
+		return (int)_employees.size();
 	}
 
 	void setData(const Employee& employee)
